Commit prior to fixing the P_00. I decided on Nov 18th 2019 that I will no longer change P_00 to be at the cloud edge but leave it at 1e5. This was done to streamline the creating of backgrounds and solutions
Now we can allow cm1 to define a background based on surface conditions
and read the P_0 and T_0 from the cm1 background at the cloud edge
height. Due to this change however cm1 and the MBM are working with different
definations of \theta in the dry case. This is related to using two different
reference pressures. \theta_l = \theta*(P_0/P_00)^(Rd/cpd) FOR DRY. The liquid
water correction also remains the same as a multiplicative factor as well.