Dimensional boussinesq solution has been added as a run case for quite some time and is stored in iinit = 1002.
iinit = 1002 has SMR issues with its initialization of the theta
perturbation. The problem was that the background liquid water
(full physics) and perturbation liquid water (Boussinesq) were
computed differently such that in the presence of no wave they
did not cancel contributing to an artifical increasing profile
of potential temperature in the cloud layer.

iinit = 1003 contains a fix for this issue but now chooses clear
and cloud conditions based on the defined location of zStar not a
0 contour of a liquid water function. In this case the Boussinesq
flat cloud edge location z0 is not the same as the full physics
flat cloud edge location. This leads to a set of discrete mesh nodes
where the water content in particular is inconsistent with the
background defination.

iinit = 1004 contains a very quick and dirty fix for the issue of
iinit = 1003 but in a quick test does not demonstrate much improvement
over the iinit = 1002 case. This is somewhat curious as I would have
expected to see bennifits from fixing the issue of 1002. This however
is not worth spending time on. Correctly fixing the issue of 1003 is
a better use of time and then a deeper dive into bennifits would be
better.

The correct fix for 1003 is to use the cloud generated background in
cm1 to determine the z0 location for the Boussinesq solution. The
principle here being to fit our Boussinesq solution to the full physics
enviroment generated by cm1. This will then inform the the P0 and
T0 values needed for the Boussinesq solution and all other remaining
solution parameters. The corresponding notes for this work can be found
on and around pg. 57 of BenchmarkDesigns-Paper.