Added isnd = 1003 to base.F
This case was added to address the cloud layer forming in different
locations. This gives slight improvements over isnd = 1001. This case
now uses the saturation mixing ratio previously calculated at the
height of itteration for the initial guess in the cloud layer as
opposed to the value calculatied at the previous height. The
itteration for saturations and liquid water appears to be a bit
sensitive to creating liquid water close to the cloud edge. In
particular for h = 625,312.5,156.25 the cloud layer is correctly
placed discretely (point before h is 0 point after h is nonzero) see
pg 58 of BenchmarkDesigns-Paper notes. For h = 1250, 2500, standard
the cloud layer is still off by a gridpoint toward the ground.

In theory the background calculation should be the same for each
height scale however in pratice I'm observing this is not the
case when it comes to cloud edge location. Although there does appear
to be very good agreement between other fields. I suspect the cloud
edge to be sensitive due to it being greater than or less than
0. Something as simple as round off error could push it into the
postive creating liquid. Computationally I do expect these to be
slightly different however the location of the cloud layer should be
made robust for benchmarking purposes.

A possible fix for this would be to integrate hydrostatic pressure
from the z0 location keeping the P and T from the z0 location as
initial data. Right now I am integrating from 0 with the same surface
temperature and pressure as the critical saturation. This in theory
leads to different conditions at z0 which may or may not be saturated
coditions. This however does not explain the dicrepancy between height
scales currently. Could result in a more robust cloud layer location
however. See BenchmarkDesigns-Paper pg. 60 for comments.